requirements.txt.txt
requests==2.32.*
python-dotenv==1.2.*